Fixes #3: Fix pushing pages without a module
Pages without a module were written with a filename that did not start
with an underscore which means you could not use the directory flag to
push all pages outside of a module. This fix will do everything in the
directory by default, with the added module filter is available.