fix: since flag parsing issues
Addresses the code & technical issues in #182 - the docs issues still remain, see details below

The approach taken attempted to minimize any breaking changes, however there are some.  As mentioned in #182, the docs do not match the code so the approach is a blend of both, attempting to align with the existing docs as much as possible.  For example, the docs mention any golang time format, but the code did not support all formats.  Additionally, the code allowed for formats not in golang layouts so the approach blends these two.

Breaking Changes
1. Duration - `Phrasal` support (e.g., 3 days) has been removed and is no longer supported.
2. Year/Week/Day - The docs did not mention how these are interpreted, however the code appears to use a "shift" approach meaning that 2024/03/01 would become 2023/03/1 if `1y` was specified.  As there were no tests, it's difficult to say for certain exactly what the previous code did for all year/day values (week was skipped entirely in the code).  This commit always treats Year as 365 days (leap year is not accounted for), Week as 7 days and Day as 24 hours so for `1y`, 2024/03/01 would become 2024/03/02

Non-Breaking (as it relates to what the previous code actually did) Changes:
1. `w` (weeks) is now supported per what the docs indicate (the code did not support it)
2. the prior code took a shotgun approach to parsing and included some formats beyond golang layouts (e.g. ).  The code in this commit takes a loose approach while parsing which accounts for the formats that the code had included and various other formats.
3. The performance should be improved over prior version
4. all golang layouts except for those that that contain timezone abbreviations are supported - the current docs indicate all are supported, however the previous code did not support many of them including those with tz abbreviations.

Required doc updates - As mentioned in #182, the docs/code did not match.  The docs should be updated as follows:
1. Time Format - Any golang layout except for those that contain timezone abbreviations (UnixDate, RFC822, RFC850, RFC1123) is supported
2. Definition of Year/Week/Day - 365 days, 7 days, 24 hours - does not account for leap year or leap sec, etc.
3. Duration Format - Only abbreviated durations are supported, `Phrasal` is no longer supported.  The valid identifiers are those defined by golang duration plus `y/m/d`
4. This isn't a change from existing code but the docs should indicate that all times are interpreted as local timezone unless a timezone value is specified

Additional Information - The approach to allow for pretty much any input when it comes to times/durations is not necessarily the best approach to take due to the complexities of parsing both especially since they share the same flag.  There is a reason that golang stdlib doesn't support `y/w/d` for duration as there are simply too many unknown ways to interpret the data and its situation dependent.  Additionally, by supporting all time formats, it leads to potential false positives.  Given the use case for the CLI where users either are unlikely to input a `--since` flag or if they do, it would most likely be a duration over a time (just guessing here), limiting the time formats supported (e.g., RFC3339 would cover all needed values) is likely best.  Even for duration, supporting `y/w/d` may be unnecessary.  This is all implemented now but simplifying the approach likely makes sense and would reduce a lot of the code complexity and maintenance cost for a flag that is likely not used very often (adding telemetry to the CLI so Skuid could observe how the CLI is used would be ideal).