fix: `since` usage
By default, pflag will generate usage text that includes the type of the flag.  In the case of `since` which is a time.Time, the usage becomes `-s, --since time.Time ....` which is rather meaningless to a user.  There is discussion (https://github.com/search?q=repo%3Aspf13%2Fpflag+unquoteusage&type=issues) in pflag repo regarding providing more control over the UsageType that is generated, however other than backticks, there is no way to customize this currently.
Modifying the usage for `since` as follows:
1. Use backticks around `time` so that it will be used for the type
2. Decrease the length of the message itself as the previous length was rather long.  The [CLI Docs]() can/should contain further and accurate details regarding the formats of time & timespan supported (see a63d27d60363b74381ce623dabbb7b30da20be86).