fix: parse entity files to strongly typed structures
The 0.67.0 version of the CLI contains a lot of string parsing related to file paths which had numerous problems as described in the issues submitted to the repo.  Prior commits to address those issues were successful and consolidated the parsing to a single location.  However, during runtime, there were still locations in the code that needed to identify things like "what metadata type" is this file and to do so, it would simply check the beginnging of the file path for a "known" metadata name.  The challenge with that is that it leads to bugs like #200 where the parsing is not accurate.

This commit provides the ability to address all remaining issues in a reliable manner by moving to strongly typed structures.  Once a path is parsed, no further parsing and/or inspection of a string is ever required as the type has all the information needed.  In short, a filepath or entity name is parsed in one place and then structs passed around (no more strings, yay!).

Beyond the obvious benefits that this yields, it provides the ability to modify the CLI to account for the many issues that exist in the server side API. In this commit, the filename/entity name is parsed and validated based on the known set of rules for each metadata type.  In theory, while the parsing to strongly typed should have existed anyway, the validation should not be needed.  However, in many situations, when the server API encounters something it doesn't "expect" in a request payload, it just ignores it and moves on rather than return an HTTP 400 Bad Request as it should.  This leaves the CLI in a vulnerable state because it may have a file locally that isn't a valid metadata file (e.g., pages/my_page.txt), send it to the server to apply and the server returns 200 but doesn't do anything.  Currently in many of these situations, the CLI doesn't emit any message regarding the unexpected result so the user thinks all was good when it really wasn't.  Having these validations client side enables the CLI to try, when possible, to avoid calling the API when we detect invalid files or at least try to be able to somehat intelligently compare what we expect to what the server said it did to log when things didn't go as expected.  This won't necessarily result in unexpected state of local and/or server metadata but in many of these situations, at least the user will know that something didn't go as expected.

In short, the server issues need to be addressed regardless of the validations that the CLI is performing client side so essentially the validation in the CLI is duplicating what needs to occur server side anyway and in theory, could be removed once all the server issues are resolved.

Resolves #200
Enables ability to add #157
Enables ability to improve validation and/or detection for issues such as #162, #163, #164, #184, #198