fix: validate archive and files being archived
Note that this is more of a workaround then a solution as the proper solution should come server-side in the API where it should be validating payload prior to executing and returning an appropriate 400 if the payload is not valid (e.g., missing required files, etc.). That said, this approach will detect the known situations that would cause the server to "ignore" an entity in the getdeployplan request so that the operation halts prior to deploying an unexpected and inconsistent payload.

Marking issues below as related rather than resolved since server-side API validation needs to be completed before these issues can be fully closed.

Related #163
Related #164
Related #184