fix: validate archive entities vs plan entities vs result entities
Note that this builds on 83f87334e8a2b26dc93b089036b5f3cf807297b0 and is more of an additional workaround then a solution.  The solution should come server-side in the API(s) where it should be validating entire payload prior to executing and returning 400 if payload is not valid (e.g., missing required files).  That said, this approach will do its best to detect when the local files archived are not in the deployment plan generated and, in turn, when the results from the deploy do match what was in the plan.  The plan vs. results is incredibly fragile since the results returned are not reliable (see #211).

Marking issues below as related rather than resolved since server-side API validation needs to be ompleted before these issues can be fully closed.

Related #163
Related #164
Related #184