test: add CI workflow with cross-platform test coverage
Note that this only targets build/test from a CI perspective and does not target code coverage.  Depending on what the goals of test coverage are for the CLI, a separate code quality/coverage workflow should be created since it likely doesn't need to run on every push or this file can be modified to include if desired.

Resolves #169
Related #166