feat: add `--no-clean` flag to retrieve
I considered a few different options here including naming the flag `overwrite` instead of `no-clean` or making it an enum with `replace` | `overwrite`.  Also considered changing the default behavior to not `clean`.  In the end, the only real way to know what the default should be is to know more about the "average" user persona and their use cases so for now, just leaving the existing default to avoid any breaking changes and adding a `no-clean` option since `replace/overwrite` could be a little ambiguous.

This is easy enough to change so input is welcomed and happy to make adjustments.

Resolves #217