fix/chore: improve logging & validations, fix several issues, refactor codebase
This commit started with the goal of improving logging (#220) to help aid with identifying when outcomes were not as expected due to the many issues currently present in the CLI and on the server-side.  During the course of improving the logging which required going through essentially every line of code in the codebase, many new issues were identified and fixes for nearly all of them implemented.  Additionally,validations of all steps in the CLI were improved to and logs emitted with the results.  Lastly, the code was significnatly refactored in many areas to improve code maintainability (#230).

Due to the size of this commit, there is too much to explain in detail in a commit message.  See forthcoming comment in #205 for full details of this commit.

Resolves #59
Resolves #165
Resolves #199
Resolves #218
Resolves #220
Resolves #221
Resolves #222
Resolves #223
Resolves #224
Resolves #228
Resolves #230
Resolves #231